@@ -1,12 +1,12 @@
 jQuery(function($){
-   $("#adopter_phone").mask("(999) 999-9999");
-   $("#adopter_other_phone").mask("(999) 999-9999");
-   $("#adopter_adoption_app_attributes_landlord_phone").mask("(999) 999-9999");
-   $("#adopter_adoption_app_attributes_vet_phone").mask("(999) 999-9999");
-   $("#user_phone").mask("(999) 999-9999");
-   $("#user_other_phone").mask("(999) 999-9999");
-   $("#adopter_references_attributes_0_phone").mask("(999) 999-9999");
-   $("#adopter_references_attributes_1_phone").mask("(999) 999-9999");
-   $("#adopter_references_attributes_2_phone").mask("(999) 999-9999");
-   $("#event_location_phone").mask("(999) 999-9999");
+   $("#adopter_phone").inputmask("(999) 999-9999");
+   $("#adopter_other_phone").inputmask("(999) 999-9999");
+   $("#adopter_adoption_app_attributes_landlord_phone").inputmask("(999) 999-9999");
+   $("#adopter_adoption_app_attributes_vet_phone").inputmask("(999) 999-9999");
+   $("#user_phone").inputmask("(999) 999-9999");
+   $("#user_other_phone").inputmask("(999) 999-9999");
+   $("#adopter_references_attributes_0_phone").inputmask("(999) 999-9999");
+   $("#adopter_references_attributes_1_phone").inputmask("(999) 999-9999");
+   $("#adopter_references_attributes_2_phone").inputmask("(999) 999-9999");
+   $("#event_location_phone").inputmask("(999) 999-9999");
 });
